Add "Push to PWS" metadata support
In order to display a "Push to PWS" button on the guides pages, we need
to check for metadata in the guide repository.
This allows us to:

* check whether we should display a button for this guide
* gather metadata to generate the button code snippet

The metadata is stored in a `"push-to-pws/button.yml"` file in the
repository, and it contains useful information for the button
generation, such as: the repository git URL, the directory that contains
the application to deploy, and the request path to call when the
application has started.