Remove Cloud custom Cloud connectors
This commit adapts the configuration for the Redis cache and the
database connection since we've moved from Cloud Connectors to Java
CFEnv for Spring Boot. Configuration properties are automatically parsed
from VCAP environment properties and contributed to the Spring Boot
configuration properties.

The Concourse pipeline needs to be updated to disable Java Buildpack
auto-reconfiguration as a result.